Tidied up code, refactored, add comments in World.java

+++ src/Item.java
@@ -44,7 +44,7 @@ public class Item {
         this.type = type;
     }
 
-    public float getDecay() {
+    public int getDecay() {
         return decay;
     }
